Add connectorRelevant Instrumented test to gather API
