+++ main/src/androidTest/java/cgeo/geocaching/filters/ConnectorRelevantFilterTest.java
@@ -1,178 +1,109 @@
 package cgeo.geocaching.filters;
 
+import cgeo.geocaching.SearchResult;
 import cgeo.geocaching.connector.IConnector;
 import cgeo.geocaching.connector.gc.GCConnector;
-import cgeo.geocaching.connector.su.SuConnector;
+import cgeo.geocaching.connector.oc.OCDEConnector;
 import cgeo.geocaching.filters.core.AndGeocacheFilter;
-import cgeo.geocaching.filters.core.BaseGeocacheFilter;
 import cgeo.geocaching.filters.core.GeocacheFilter;
 import cgeo.geocaching.filters.core.GeocacheFilterType;
-import cgeo.geocaching.filters.core.NameGeocacheFilter;
+import cgeo.geocaching.filters.core.IGeocacheFilter;
 import cgeo.geocaching.filters.core.NotGeocacheFilter;
 import cgeo.geocaching.filters.core.OrGeocacheFilter;
 import cgeo.geocaching.filters.core.OriginGeocacheFilter;
+import cgeo.geocaching.filters.core.OwnerGeocacheFilter;
+import cgeo.geocaching.models.Geocache;
+import cgeo.geocaching.sorting.GeocacheSort;
 
-import java.util.List;
+import java.util.Collections;
+import java.util.Set;
 
 import org.junit.Test;
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
 /**
- * Tests for {@link GeocacheFilter#getConnectorRelevantFilter(IConnector)} and
- * {@link GeocacheFilter#getAndChainIfPossible(IConnector)}.
+ * Instrumented API test for the combination of OriginGeocacheFilter and OwnerGeocacheFilter.
+ * Tests OR(AND(Origin=GC, Owner="lineflyer"), AND(Origin=OC, Owner="mic@")) against the
+ * real geocaching.com and opencaching.de APIs.
+ * <p>
+ * Requires configured accounts for geocaching.com and opencaching.de on the emulator/device.
  */
 public class ConnectorRelevantFilterTest {
-
-    private static final IConnector GC = GCConnector.getInstance();
-    private static final IConnector SU = SuConnector.getInstance();
-
-    // --- Tests ---
+    private static final GCConnector GC = GCConnector.getInstance();
+    private static final OCDEConnector OC = new OCDEConnector();
+    private static final String GC_TEST_OWNER = "Grokky Grokson";
+    private static final String OC_TEST_OWNER = "lineflyer";
 
     @Test
-    public void nullTreeReturnsEmptyFilter() {
-        final GeocacheFilter empty = GeocacheFilter.createEmpty();
-        final GeocacheFilter result = empty.getConnectorRelevantFilter(GC);
-        assertThat(result).isNotNull();
-        assertThat(result.getTree()).isNull();
-        assertThat(result.isFiltering()).isFalse();
+    public void searchGCOrOCFilter() {
+        // OR(AND(Origin=GC, Owner=GC_TEST_OWNER), AND(Origin=OC, Owner=OC_TEST_OWNER))
+        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, createGCOrOCFilter());
+
+        // Sent to GC API: only GC branch is relevant → Owner=GC_TEST_OWNER
+        final SearchResult gcSearch = GC.searchByFilter(filter, new GeocacheSort());
+        assertThat(gcSearch).isNotNull();
+        assertThat(gcSearch.getGeocodes()).isNotEmpty();
+        assertOwner(gcSearch, GC_TEST_OWNER);
+
+        // Sent to OC API: only OC branch is relevant → Owner=OC_TEST_OWNER
+        final SearchResult ocSearch = OC.searchByFilter(filter, new GeocacheSort());
+        assertThat(ocSearch).isNotNull();
+        assertThat(ocSearch.getGeocodes()).isNotEmpty();
+        assertOwner(ocSearch, OC_TEST_OWNER);
     }
 
     @Test
-    public void simpleAndWithMatchingOriginReturnsFilterWithoutOrigin() {
-        // AND(Origin=GC, Name="test")
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false,
-                AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("test")));
+    public void searchIgnoreOCFilter() {
+        // AND(NOT(Origin=OC), Owner=OC_TEST_OWNER))
+        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, createIgnoreOCFilter());
 
-        final GeocacheFilter result = filter.getConnectorRelevantFilter(GC);
-        assertThat(result).isNotNull();
-        // Origin was removed, only NameFilter remains
-        assertThat(result.getTree()).isInstanceOf(NameGeocacheFilter.class);
+        // Sent to GC API: only GC branch is relevant → Owner=OC_TEST_OWNER
+        final SearchResult gcSearch = GC.searchByFilter(filter, new GeocacheSort());
+        assertThat(gcSearch).isNotNull();
+        assertThat(gcSearch.getGeocodes()).isNotEmpty();
+        assertOwner(gcSearch, OC_TEST_OWNER);
+
+        // Sent to OC API: OC branch is not relevant → empty SearchResult
+        final SearchResult ocSearch = OC.searchByFilter(filter, new GeocacheSort());
+        assertThat(ocSearch).isNotNull();
+        assertThat(ocSearch.getGeocodes()).isEmpty();
     }
 
-    @Test
-    public void simpleAndWithNonMatchingOriginReturnsNull() {
-        // AND(Origin=GC, Name="test")
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false,
-                AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("test")));
-
-        final GeocacheFilter result = filter.getConnectorRelevantFilter(SU);
-        assertThat(result).isNull();
+    private static void assertOwner(SearchResult search, final String expectedOwner) {
+        // Verify all caches belong to the expected owner
+        final Set<Geocache> caches = search.getCachesFromSearchResult();
+        assertThat(caches).isNotEmpty();
+        for (final Geocache cache : caches) {
+            assertThat(cache.getOwnerDisplayName())
+                    .as("Owner of cache " + cache.getGeocode())
+                    .isEqualToIgnoringCase(expectedOwner);
+        }
     }
 
-    @Test
-    public void orWithDifferentOriginBranches() {
-        // OR(AND(Origin=GC, Name="gc-test"), AND(Origin=SU, Name="su-test"))
-        final AndGeocacheFilter gcBranch = AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("gc-test"));
-        final AndGeocacheFilter suBranch = AndGeocacheFilter.create(OriginGeocacheFilter.create(SU), NameGeocacheFilter.create("su-test"));
-        final OrGeocacheFilter orFilter = new OrGeocacheFilter();
-        orFilter.addChild(gcBranch);
-        orFilter.addChild(suBranch);
-
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, orFilter);
-
-        // For GC: only gc-branch should remain
-        final GeocacheFilter gcResult = filter.getConnectorRelevantFilter(GC);
-        assertThat(gcResult).isNotNull();
-        assertThat(gcResult.getTree()).isInstanceOf(NameGeocacheFilter.class);
-
-        // For SU: only su-branch should remain
-        final GeocacheFilter suResult = filter.getConnectorRelevantFilter(SU);
-        assertThat(suResult).isNotNull();
-        assertThat(suResult.getTree()).isInstanceOf(NameGeocacheFilter.class);
+    private static IGeocacheFilter createGCOrOCFilter() {
+        // OR(AND(Origin=GC, Owner=GC_TEST_OWNER), AND(Origin=OC, Owner=OC_TEST_OWNER))
+        return OrGeocacheFilter.create(createOwnerBranch(GC, GC_TEST_OWNER), createOwnerBranch(OC, OC_TEST_OWNER));
     }
 
-    @Test
-    public void noOriginFilterReturnsFullTree() {
-        // AND(Name="test")
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, NameGeocacheFilter.create("test"));
-
-        final GeocacheFilter gcResult = filter.getConnectorRelevantFilter(GC);
-        assertThat(gcResult).isNotNull();
-        assertThat(gcResult.getTree()).isInstanceOf(NameGeocacheFilter.class);
-
-        final GeocacheFilter suResult = filter.getConnectorRelevantFilter(SU);
-        assertThat(suResult).isNotNull();
-        assertThat(suResult.getTree()).isInstanceOf(NameGeocacheFilter.class);
+    private static IGeocacheFilter createIgnoreOCFilter() {
+        // AND(NOT(Origin=OC), Owner=OC_TEST_OWNER))
+        return AndGeocacheFilter.create(NotGeocacheFilter.create(createOriginFilter(OC)), createOwnerFilter(OC_TEST_OWNER));
     }
 
-    @Test
-    public void notOriginExcludesMatchingConnector() {
-        // NOT(Origin=GC) means: exclude GC caches
-        final NotGeocacheFilter notFilter = new NotGeocacheFilter();
-        notFilter.addChild(OriginGeocacheFilter.create(GC));
-
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, notFilter);
-
-        // GC should be excluded (NOT(always-true-for-GC) = false)
-        assertThat(filter.getConnectorRelevantFilter(GC)).isNull();
-
-        // SU should NOT be excluded (NOT(false-for-SU) = true → always true)
-        final GeocacheFilter suResult = filter.getConnectorRelevantFilter(SU);
-        assertThat(suResult).isNotNull();
-        assertThat(suResult.getTree()).isNull(); // simplified to always-true
+    private static IGeocacheFilter createOwnerBranch(final IConnector connector, final String owner) {
+        return AndGeocacheFilter.create(createOriginFilter(connector), createOwnerFilter(owner));
     }
 
-    @Test
-    public void nonFilteringOriginIsAlwaysTrue() {
-        // empty OriginFilter (isFiltering() == false) → always true
-        final OriginGeocacheFilter emptyOrigin = GeocacheFilterType.ORIGIN.create();
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, emptyOrigin);
-
-        final GeocacheFilter result = filter.getConnectorRelevantFilter(GC);
-        assertThat(result).isNotNull();
-        assertThat(result.getTree()).isNull(); // simplified to always-true
+    private static IGeocacheFilter createOriginFilter(final IConnector connector) {
+        final OriginGeocacheFilter originFilter = GeocacheFilterType.ORIGIN.create();
+        originFilter.setValues(Collections.singletonList(connector));
+        return originFilter;
     }
 
-    @Test
-    public void getAndChainIfPossibleWithConnectorReturnsNullForExcluded() {
-        // AND(Origin=GC, Name="test")
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false,
-                AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("test")));
-
-        final List<BaseGeocacheFilter> gcChain = filter.getAndChainIfPossible(GC);
-        assertThat(gcChain).isNotNull();
-        assertThat(gcChain).hasSize(1);
-        assertThat(gcChain.get(0)).isInstanceOf(NameGeocacheFilter.class);
-
-        final List<BaseGeocacheFilter> suChain = filter.getAndChainIfPossible(SU);
-        assertThat(suChain).isNull();
+    private static IGeocacheFilter createOwnerFilter(final String owner) {
+        final OwnerGeocacheFilter ownerFilter = GeocacheFilterType.OWNER.create();
+        ownerFilter.getStringFilter().setTextValue(owner);
+        return ownerFilter;
     }
 
-    @Test
-    public void orWithMixedBranchesPartiallyRelevant() {
-        // OR(AND(Origin=GC, Name="gc-test"), Name="general")
-        final AndGeocacheFilter gcBranch = AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("gc-test"));
-        final NameGeocacheFilter generalBranch = NameGeocacheFilter.create("general");
-        final OrGeocacheFilter orFilter = new OrGeocacheFilter();
-        orFilter.addChild(gcBranch);
-        orFilter.addChild(generalBranch);
-
-        final GeocacheFilter filter = GeocacheFilter.create(null, false, false, orFilter);
-
-        // For GC: both branches are relevant → OR(Name="gc-test", Name="general")
-        final GeocacheFilter gcResult = filter.getConnectorRelevantFilter(GC);
-        assertThat(gcResult).isNotNull();
-        assertThat(gcResult.getTree()).isInstanceOf(OrGeocacheFilter.class);
-        assertThat(((OrGeocacheFilter) gcResult.getTree()).getChildren()).hasSize(2);
-
-        // For SU: only the general branch is relevant → Name="general"
-        final GeocacheFilter suResult = filter.getConnectorRelevantFilter(SU);
-        assertThat(suResult).isNotNull();
-        assertThat(suResult.getTree()).isInstanceOf(NameGeocacheFilter.class);
-    }
-
-    @Test
-    public void preservesNameAndFlags() {
-        final GeocacheFilter filter = GeocacheFilter.create("myFilter", true, true,
-                AndGeocacheFilter.create(OriginGeocacheFilter.create(GC), NameGeocacheFilter.create("test")));
-
-        final GeocacheFilter result = filter.getConnectorRelevantFilter(GC);
-        assertThat(result).isNotNull();
-        assertThat(result.getName()).isEqualTo("myFilter");
-        assertThat(result.isOpenInAdvancedMode()).isTrue();
-        assertThat(result.isIncludeInconclusive()).isTrue();
-    }
 }
-
-
-
